requirements.txt
fastapi==0.111.0
sqlmodel==0.0.18
pydantic==2.7.1
SQLAlchemy==2.0.30
requests==2.32.2
pydantic-settings==2.2.1
passlib==1.7.4
pyjwt==2.8.0
email_validator==2.1.1
loguru==0.7.2
psycopg2-binary=2.9.9
sqlalchemy-utils==0.41.2
alembic==1.13.2
asyncpg==0.29.0
python-multipart==0.0.9
fastapi-mail==1.4.1
itsdangerous==2.2.0
redis==5.0.8
